folder refactor: last cleanup
