+++ src/metrics/first-meaningful-paint.js
@@ -16,7 +16,7 @@
  */
 'use strict';
 
-const DevtoolsTimelineModel = require('../../lib/traces/devtools-timeline-model');
+const DevtoolsTimelineModel = require('../lib/traces/devtools-timeline-model');
 
 const FAILURE_MESSAGE = 'Navigation and first paint timings not found.';
