+++ test/audits/manifest/exists.js
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-const Audit = require('../../../audits/manifest/exists.js');
+const Audit = require('../../../src/audits/manifest/exists.js');
 const assert = require('assert');
 
 /* global describe, it*/
