+++ test/audits/manifest/short-name.js
@@ -13,10 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-const Audit = require('../../../audits/manifest/short-name.js');
+const Audit = require('../../../src/audits/manifest/short-name.js');
 const assert = require('assert');
 const manifestSrc = JSON.stringify(require('./manifest.json'));
-const manifestParser = require('../../../lib/manifest-parser');
+const manifestParser = require('../../../src/lib/manifest-parser');
 const manifest = manifestParser(manifestSrc);
 
 /* global describe, it*/
