+++ test/audits/manifest/start-url.js
@@ -13,10 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-const Audit = require('../../../audits/manifest/start-url.js');
+const Audit = require('../../../src/audits/manifest/start-url.js');
 const assert = require('assert');
 const manifestSrc = JSON.stringify(require('./manifest.json'));
-const manifestParser = require('../../../lib/manifest-parser');
+const manifestParser = require('../../../src/lib/manifest-parser');
 const manifest = manifestParser(manifestSrc);
 
 /* global describe, it*/
