+++ test/audits/mobile-friendly/viewport.js
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-const Audit = require('../../../audits/mobile-friendly/viewport.js');
+const Audit = require('../../../src/audits/mobile-friendly/viewport.js');
 const assert = require('assert');
 
 /* global describe, it*/
