+++ test/audits/security/is-on-https.js
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-const Audit = require('../../../audits/security/is-on-https.js');
+const Audit = require('../../../src/audits/security/is-on-https.js');
 const assert = require('assert');
 
 /* global describe, it*/
